text-to-image/requirements.txt
diffusers
transformers
accelerate
torch
safetensors
huggingface_hub
gradio
python-dotenv
